# docs/swagger.yaml
basePath: /
definitions:
...
  database.CreateVehicleParams:
    properties:
      capacity:
        items:
          type: integer
        type: array
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      end_location:
        $ref: '#/definitions/util.LocationParams'
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
      speed_factor:
        type: number
      start_location:
        $ref: '#/definitions/util.LocationParams'
      tw_close:
        type: string
      tw_open:
        type: string
    required:
    - end_location
    - start_location
    type: object
...
  database.UpdateVehicleParams:
    properties:
      capacity:
        items:
          type: integer
        type: array
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      end_location:
        $ref: '#/definitions/util.LocationParams'
      project_id:
        example: "1234567812345678"
        type: string
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
      speed_factor:
        type: number
      start_location:
        $ref: '#/definitions/util.LocationParams'
      tw_close:
        type: string
      tw_open:
        type: string
    type: object
  database.Vehicle:
    properties:
      capacity:
        items:
          type: integer
        type: array
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      end_location:
        $ref: '#/definitions/util.LocationParams'
      id:
        example: "1234567812345678"
        type: string
      project_id:
        example: "1234567812345678"
        type: string
      skills:
        example:
        - 1
        - 5
        items:
          type: integer
        type: array
      speed_factor:
        type: number
      start_location:
        $ref: '#/definitions/util.LocationParams'
      tw_close:
        example: "2021-12-31 23:59:00"
        type: string
      tw_open:
        example: "2021-12-31 23:00:00"
        type: string
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
  util.LocationParams:
    properties:
      latitude:
        example: 2.0365
        type: number
      longitude:
        example: 48.6113
        type: number
    required:
    - latitude
    - longitude
    type: object
  util.MultiError:
    properties:
      errors:
        example:
        - Error message1
        - Error message2
        items:
          type: string
        type: array
    type: object
  util.NotFound:
    properties:
      error:
        example: Not Found
        type: string
    type: object
...
  util.Success:
    properties:
      success:
        example: "true"
        type: string
    type: object
host: localhost:9100
info:
  contact:
    email: info@georepublic.de
    name: Team Georepublic
  description: This is an API for scheduling VRP tasks. Source code can be found on
    https://github.com/Georepublic/pg_scheduleserv
  license:
    name: GNU Affero General Public License
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://swagger.io/terms/
  title: pg_scheduleserv API
  version: 0.1.0
paths:
...
  /projects/{project_id}/vehicles:
    get:
      consumes:
      - application/json
      description: Get a list of vehicles for a project with project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List vehicles for a project
      tags:
      - Vehicle
    post:
      consumes:
      - application/json
      description: Create a new vehicle with the input payload
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Create vehicle
        in: body
        name: Vehicle
        required: true
        schema:
          $ref: '#/definitions/database.CreateVehicleParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new vehicle
      tags:
      - Vehicle
...
  /vehicles/{vehicle_id}:
    delete:
      consumes:
      - application/json
      description: Delete a vehicle with its vehicle_id
      parameters:
      - description: Vehicle ID
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete a vehicle
      tags:
      - Vehicle
    get:
      consumes:
      - application/json
      description: Fetch a vehicle with its vehicle_id
      parameters:
      - description: Vehicle ID
        in: path
        name: vehicle_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.NotFound'
      summary: Fetch a vehicle
      tags:
      - Vehicle
    patch:
      consumes:
      - application/json
      description: Update a vehicle with its vehicle_id
      parameters:
      - description: Vehicle ID
        in: path
        name: vehicle_id
        required: true
        type: integer
      - description: Update vehicle
        in: body
        name: Vehicle
        required: true
        schema:
          $ref: '#/definitions/database.UpdateVehicleParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Vehicle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Update a vehicle
      tags:
      - Vehicle
...
schemes:
- http
- https
swagger: "2.0"